requirements.txt
streamlit
streamlit-authenticator
streamlit-extras
pyyaml